+++ src/main/java/net/mohron/skyclaims/command/argument/BiomeArgument.java
@@ -21,7 +21,6 @@ package net.mohron.skyclaims.command.argument;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import net.mohron.skyclaims.SkyClaims;
-import net.mohron.skyclaims.config.type.PermissionConfig;
 import net.mohron.skyclaims.permissions.Permissions;
 import org.spongepowered.api.command.CommandSource;
 import org.spongepowered.api.command.args.ArgumentParseException;