+++ src/main/java/net/mohron/skyclaims/config/type/GlobalConfig.java
@@ -24,27 +24,19 @@ import ninja.leaping.configurate.objectmapping.serialize.ConfigSerializable;
 
 @ConfigSerializable
 public class GlobalConfig {
-	@Setting(value = "Config-Version")
-	private int version;
-	@Setting(value = "Permission")
-	private PermissionConfig permissionConfig;
-	@Setting(value = "Misc")
-	private MiscConfig miscConfig;
-	@Setting(value = "Storage")
-	private StorageConfig storageConfig;
-	@Setting(value = "World")
-	private WorldConfig worldConfig;
-	@Setting(value = "Options", comment = "The default values options use when not explicitly set.\ngithub.com/DevOnTheRocks/SkyClaims/wiki/Options")
-	private OptionsConfig optionsConfig;
 
-	public GlobalConfig() {
-		version = ConfigManager.CONFIG_VERSION;
-		permissionConfig = new PermissionConfig();
-		miscConfig = new MiscConfig();
-		storageConfig = new StorageConfig();
-		worldConfig = new WorldConfig();
-		optionsConfig = new OptionsConfig();
-	}
+	@Setting(value = "Config-Version")
+	private int version = ConfigManager.CONFIG_VERSION;
+	@Setting(value = "Permission")
+	private PermissionConfig permissionConfig = new PermissionConfig();
+	@Setting(value = "Misc")
+	private MiscConfig miscConfig = new MiscConfig();
+	@Setting(value = "Storage")
+	private StorageConfig storageConfig = new StorageConfig();
+	@Setting(value = "World")
+	private WorldConfig worldConfig = new WorldConfig();
+	@Setting(value = "Options", comment = "The default values options use when not explicitly set.\ngithub.com/DevOnTheRocks/SkyClaims/wiki/Options")
+	private OptionsConfig optionsConfig = new OptionsConfig();
 
 	public int getVersion() {
 		return version;