+++ src/main/java/net/mohron/skyclaims/config/type/MysqlConfig.java
@@ -23,27 +23,19 @@ import ninja.leaping.configurate.objectmapping.serialize.ConfigSerializable;
 
 @ConfigSerializable
 public class MysqlConfig {
-	@Setting("Name")
-	private String databaseName;
-	@Setting("Location")
-	private String location;
-	@Setting("Table-Prefix")
-	private String tablePrefix;
-	@Setting("Username")
-	private String username;
-	@Setting("Password")
-	private String password;
-	@Setting("Port")
-	private Integer port;
 
-	public MysqlConfig() {
-		databaseName = "skyclaims";
-		location = "localhost";
-		tablePrefix = "";
-		username = "skyclaims";
-		password = "skyclaims";
-		port = 3306;
-	}
+	@Setting("Name")
+	private String databaseName = "skyclaims";
+	@Setting("Location")
+	private String location = "localhost";
+	@Setting("Table-Prefix")
+	private String tablePrefix = "";
+	@Setting("Username")
+	private String username = "skyclaims";
+	@Setting("Password")
+	private String password = "skyclaims";
+	@Setting("Port")
+	private int port = 3306;
 
 	public String getDatabaseName() {
 		return databaseName;
@@ -66,6 +58,6 @@ public class MysqlConfig {
 	}
 
 	public int getPort() {
-		return (port != null) ? port : 3306;
+		return port;
 	}
 }