+++ src/main/java/net/mohron/skyclaims/config/type/PermissionConfig.java
@@ -23,18 +23,13 @@ import ninja.leaping.configurate.objectmapping.serialize.ConfigSerializable;
 
 @ConfigSerializable
 public class PermissionConfig {
-	@Setting(value = "Separate-BiomeType-Permissions", comment = "Enable permission checking for the Biome Type Argument.")
-	private boolean separateBiomePerms;
-	@Setting(value = "Separate-Schematic-Permissions", comment = "Enable permission checking for the Schematic Argument.")
-	private boolean separateSchematicPerms;
-	@Setting(value = "Separate-Target-Permissions", comment = "Enable permission checking for the Target Argument.")
-	private boolean separateTargetPerms;
 
-	public PermissionConfig() {
-		separateBiomePerms = false;
-		separateSchematicPerms = false;
-		separateTargetPerms = false;
-	}
+	@Setting(value = "Separate-BiomeType-Permissions", comment = "Enable permission checking for the Biome Type Argument.")
+	private boolean separateBiomePerms = false;
+	@Setting(value = "Separate-Schematic-Permissions", comment = "Enable permission checking for the Schematic Argument.")
+	private boolean separateSchematicPerms = false;
+	@Setting(value = "Separate-Target-Permissions", comment = "Enable permission checking for the Target Argument.")
+	private boolean separateTargetPerms = false;
 
 	public boolean isSeparateBiomePerms() {
 		return separateBiomePerms;