+++ src/main/java/net/mohron/skyclaims/config/type/StorageType.java
@@ -16,15 +16,8 @@
  * along with SkyClaims.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package net.mohron.skyclaims;
+package net.mohron.skyclaims.config.type;
 
-public class PluginInfo {
-	public static final String ID = "skyclaims";
-	public static final String NAME = "@NAME@";
-	public static final String VERSION = "@VERSION@";
-	public static final String DESCRIPTION = "@DESCRIPTION@";
-	public static final String AUTHORS = "Mohron, Cossacksman";
-	public static final String GP_VERSION = "2.3.1";
-	public static final double GP_API_VERSION = 0.2;
-	public static final String NUCLEUS_VERSION = "0.26.0";
+public enum StorageType {
+	SQLite, MySQL
 }