+++ src/main/java/net/mohron/skyclaims/database/Database.java
@@ -34,33 +34,31 @@ import java.util.Map;
 import java.util.UUID;
 
 public abstract class Database implements IDatabase {
+
 	abstract Connection getConnection() throws SQLException;
-	private String tablename = "island";
+
+	String tableName;
+	String primaryKey;
 
 	/***
 	 * Creates a table using SQL syntax for Sqlite or Mysql
 	 */
 	void createTable() {
-		String type = SkyClaims.getInstance().getConfig().getStorageConfig().getType();
-
-		if(type.equals("MySQL")){
-			this.tablename =  SkyClaims.getInstance().getConfig().getStorageConfig().getMysqlConfig().getTablePrefix() + "island";
-		}
 
 		try (Statement statement = getConnection().createStatement()) {
 			statement.setQueryTimeout(30);
 
 			// Create the database schema
-			String table =
-				"CREATE TABLE IF NOT EXISTS `" + this.tablename + "` ( " +
-				"`island` TEXT NOT NULL, "+
+			String table = "CREATE TABLE IF NOT EXISTS `" +
+				tableName + "` ( " +
+				"`island` TEXT NOT NULL, " +
 				"`owner` TEXT NOT NULL, " +
 				"`claim` TEXT NOT NULL, " +
 				"`spawnX` INT NOT NULL, " +
 				"`spawnY` INT NOT NULL, " +
 				"`spawnZ` INT NOT NULL, " +
 				"`locked` BOOLEAN NOT NULL, " +
-				"PRIMARY KEY " + (type.equals("SQLite") ? "(`island`)" : "(`island`(60))") + ");";
+				"PRIMARY KEY " + primaryKey + ");";
 
 			// Create the islands table (execute statement)
 			statement.executeUpdate(table);
@@ -79,7 +77,7 @@ public abstract class Database implements IDatabase {
 		HashMap<UUID, Island> islands = Maps.newHashMap();
 
 		try (Statement statement = getConnection().createStatement()) {
-			ResultSet results = statement.executeQuery("SELECT * FROM `" + this.tablename + "`" );
+			ResultSet results = statement.executeQuery("SELECT * FROM `" + tableName + "`");
 
 			while (results.next()) {
 				UUID islandId = UUID.fromString(results.getString("island"));
@@ -111,8 +109,7 @@ public abstract class Database implements IDatabase {
 	 * @param islands The collection in memory to pull the data from
 	 */
 	public void saveData(Collection<Island> islands) {
-		for (Island island : islands)
-			saveIsland(island);
+		islands.forEach(this::saveIsland);
 	}
 
 	/**
@@ -121,8 +118,7 @@ public abstract class Database implements IDatabase {
 	 * @param islands The map in memory to pull the data from
 	 */
 	public void saveData(Map<UUID, Island> islands) {
-		for (Island island : islands.values())
-			saveIsland(island);
+		islands.forEach((uuid, island) -> saveIsland(island));
 	}
 
 	/**
@@ -131,7 +127,8 @@ public abstract class Database implements IDatabase {
 	 * @param island the island to save
 	 */
 	public void saveIsland(Island island) {
-		String sql = "REPLACE INTO `" + this.tablename + "` (island, owner, claim, spawnX, spawnY, spawnZ, locked) VALUES(?, ?, ?, ?, ?, ?, ?)";
+		String sql = String.format("REPLACE INTO `%s` (island, owner, claim, spawnX, spawnY, spawnZ, locked) VALUES(?, ?, ?, ?, ?, ?, ?)",
+			this.tableName);
 
 		try (PreparedStatement statement = getConnection().prepareStatement(sql)) {
 			statement.setString(1, island.getUniqueId().toString());
@@ -154,7 +151,7 @@ public abstract class Database implements IDatabase {
 	 * @param island the island to remove
 	 */
 	public void removeIsland(Island island) {
-		String sql = "DELETE FROM `" + this.tablename + "` WHERE island = ?";
+		String sql = "DELETE FROM `" + this.tableName + "` WHERE island = ?";
 
 		try (PreparedStatement statement = getConnection().prepareStatement(sql)) {
 			statement.setString(1, island.getUniqueId().toString());
@@ -173,7 +170,7 @@ public abstract class Database implements IDatabase {
 	public int countColumns() {
 		int total = 0;
 
-		String sql = "SELECT * FROM `" + this.tablename + "` LIMIT 1";
+		String sql = "SELECT * FROM `" + this.tableName + "` LIMIT 1";
 		try (PreparedStatement statement = getConnection().prepareStatement(sql)) {
 			return statement.executeQuery().getMetaData().getColumnCount();
 		} catch (SQLException e) {