+++ src/main/java/net/mohron/skyclaims/util/ClaimUtil.java
@@ -31,7 +31,6 @@ import org.spongepowered.api.entity.living.player.User;
 import org.spongepowered.api.service.user.UserStorageService;
 import org.spongepowered.api.text.Text;
 import org.spongepowered.api.text.format.TextColors;
-import org.spongepowered.api.world.World;
 
 import java.util.List;
 import java.util.Optional;